integrate pr feedback
